[MIG] dms_field: Migration to 16.0
In the migration, the view type is changed from dms_tree to
dms_list. That's because in this point https://github.com/odoo/odoo/blob/16.0/addons/web/static/src/views/fields/field.js#L282
the 'tree' string is replaced by 'list' what is causing problems
trying to set the view type as it was before.